refactor: Provider 컴포넌트 경로 변경 및 import 경로 수정
Provider 컴포넌트를 /utils/provider.tsx로 이동했다.
또한, Provider를 가져오는 layout의 import 경로를 수정했다.
