+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import "@/styles/globals.css";
 import type { Metadata } from "next";
 import Background from "@/components/atoms/Background";
 import NavigationBar from "@/components/molecules/NavigationBar";
-import Provider from "@/components/atoms/Provider/Provider";
+import Provider from "@/utils/provider";
 
 export const metadata: Metadata = {
   title: "Thunder bowling",
